change test name to match module name
